feat/ISD-4619: Fix integration test by waiting for the HTTP after all_active wait
